Add a premises service
This fetches the list of premises from the client, and then carries out
operations on the raw data. I’ve added a `tableRows` method to return
an alphabetically ordered array of table rows for use in a view.

Currently, we’re not 100% sure how we’re going to authenticate with the
API, so we’re just throwing in a stub token for us to come back and fix
later.