Add a PremisesController and views
This adds a controller class and an index method that uses the service
to return the table rows and pass them to the view.

We can then add the PremisesController to the routes. There’s probably a
neater way to do this, but I don’t think it’s worth optimising too
prematurely.