Add bookingsController
This adds a bookingsController with two methods:
`new` renders the form to create the booking.
`create` takes the data from the form and makes a post request to the
API via the previously created service and client. `create` doesn’t do
any validation on the data, this will be handled by the API.