Add method to return a summary list for a premises
This adds a new `getPremisesDetails` method that returns data for
a particular premises for use in show template. It returns the name of
the premises and the details as an object which can be used in a
`govukSummaryList` macro. I’ve also refactored the `tableRows` method
to use a private method to return either a HMTL or Text item to DRY
things up a bit, as well as adding some new types.