Add view and E2E tests
Here we add the view to create the booking and a happy path test to
verify that the inputs exist and the user can enter details into them.
Again there is not validation for the inputs, the user could enter bad
data or no data - validation is carried out by the API and will be
passed up to the view.