Create the confirmation Page
We also update the e2e tests.
Rather than extend the already pretty length booking.cy.ts I have opted
to create a new page object and test file for the booking page.
We miss the assertion that the booking confirmation page has the
info of the booking that just posted but I think the bookingConfirmation
page tests give us enough confidence that it works as expected