Change date converter to return an ISO 8601 string
As we don’t want to do client-side validation, we tweak the
`convertDateInputsToDateObj` function to attempt to return an ISO8601
formatted datetime string. We don’t care if the string is an invalid ISO
string, as we will run the validation on the server side.

With this in mine, I’ve updated the `ObjectWithDateParts` to
have a date key that is a string, instead of a date.

I’ve also changed the bookings controller to simplify the
date conversion