Make date fields into strings
This helps keep things more predictable across the application. When we
get dates from the API, they’ll be strings anyway, so we’ll always have
to do the conversion after we return from the API anyway, so it makes
sense to be clear about the shape of the data we expect.