Add a `renderWithErrors` function
This accepts a response, request, error object, as well as the template
to render, and any additional params to pass to the template.

We then transform the error body into something that we can use in
Nunjucks templates, and pass those errors to the template, together
with the request body and additional params.

If the error is not a SantitisedError, we rethrow it.