Add nonArrival error message stubs
This also adds an additional option to the `errorStub` method to allow
us to check for null objects. This was a bit of a faff to sort, but seems to
be the only way to check for items (like radio buttons or checkboxes) that don’t
get sent when submitting without an item been selected.