Add a catchValidationErrorOrPropogate error
This works in a similar way to `renderWithErrors`, but uses flash
messages to store the errors and the user input, then redirects to the
given path. This saves us having to repeat any setup processes when
re-rendering the for with errors, or carry out any expensive
setup/teardown if the error isn’t a validation error (such as a timeout,
or something unpredictable)

I’ll replace `renderWithErrors` with this function over the next few
commits.