Use `catchValidationErrorOrPropogate` in arrivals
This uses the `catchValidationErrorOrPropogate` approach in the arrivals
controller. I’ve also tweaked the arrival form to make sure the
appropriate option is checked using the presence or otherwise of the
`arrival` / `nonArrival` object, rather than setting an arbitrary value.
This is going to be tweaked in future anyway, so both arrivals and
non-arrivals have their own views