Use new error handling in Departures
I’ve added a lot of the setup / teardown for the `new` describe block,
but had some issues with a
`Result of the Symbol.iterator method is not an object` error, which
I tracked down to the `getReferenceData` stub. Rather than spending ages
on this (seemingly obscure) error, I’ve repeated the `getReferenceData`
stub in both tests.