Add a utils method to return actions for a booking
This returns menu items for the `mojIdentityBar` depending on the state
of the booking. If the booking is awaiting arrival, we can mark it as
arrived, mark it as not arrived, or cancel it. If it’s already arrived,
we can mark it as departed. In all other cases, we return no actions.
This may be subject to change, but this will come out in testing.