Add types for Cancellations
This adds a Cancellation type and a CancellationDto type. The DTO
removes the ID and BookingID from the Cancellation type, and adds date
parts (from the GOV.UK date input), as well as changing the type of the
reason to a uuid string. When the Cancellation is returned from the DTO,
we transform the reason into a `CancellationReason` object (a subtype of
`ReferenceData`) and the date parts into a ISO8601 string.