Update E2E tests
Bringing it all together…
We run through the full E2E journey of making a booking. This makes the
first test relatively lengthy but it’s worthwhile to test the journey
in sequence.
For errors we show the CRN missing error on the ‘find’ page first before
entering a CRN and progressing to the ‘new’ page where we verify the
errors for missing fields on that page.