Bookings Controller changes
Making a booking is now a 3 step process.
1: the user is presented with a form to enter the person’s CRN.
2. We look up the person via the entry in the flash for the CRN that the
PeopleController creates. The user is presented with confirmation of
the name for the CRN they have just entered and a form to enter the rest
of the booking details.
The behaviour of this form is unchanged from the original ‘Make a booking’
form - on error the user is shown the errors
3. on success they are redirected to the confirmation screen.