Create People Controller
This will be reused in the Apply journey as well as the Manage journey
so I have made it reusable.
If the person is found (IE the Service doesn’t error) then the person’s
CRN is put in a flash which is accessed by the Booking controller and we
are redirected to the referral URL.
If any error occurs then we are sent to the referral URL and generic errors
will show on the form, this will need more work when a person can’t be found etc.