Add capacity methods to PremisesService
First we need to map the dateCapacities from the API into date ranges
that the AP is overcapacity. Then we need to generate the HTML for the
template to display when the AP is overcapacity.
I’ve opted to extract getDateRangesWithNegativeBeds to a util and unit
test it as opposed to having it as a private function in order to document
what it’s doing.
There are only 4 possible outcomes for generateOvercapacityMessage however
so I think it’s save to assume that it’s covered by the tests for getOverCapacityMessage.
The message has to be returned in an array to avoid modifying the _messages.njk
macro that shows the flash messages

Extract getDateRangeWithNegativeBeds function