Optionally load data from session or user input
When we display a page for rendering, we will want to fetch data from
the session (if we’re showing a page that’s already been completed), or
the user’s input (if we’re showing an error). This adds an optional
`userInput` param to accept the user’s input and adds a `getBody`
private method to check what type of input to pass to the page object.

If we have `userInput` provided (i.e. we’re rendering an error), we use
that, if not, we check the body of the request. Lastly, if the body is
empty, we load the session data for that page / task.

Using the page objects for the user data gives us confidence that the
data we’re returning is in the right shape without having to strongly
type the session variables.