Controller passes `userInput` to the service
Now we have an optional `userInput` argument for our service method, we
can now user this in the controller, and then pass the resulting body to
the template. This allows us to show the user’s input if there’s been
an error, or the data loaded from the session.