Create util func for retrieving data from the session
The structure of the session object is liable to change and it is better
to have to make those changes in one place than have them throughout
the code.
I did think about having this as a static method on the applicationService
but I think it’s best not to have the data flowing in both directions.