Create 'new' method in ApplicationsController
This method is almost identical to the ‘new’ method in the bookingController.
We check to see if there is a property called ‘crn’ present in the flash.
If there isn’t then we present a form for the user to enter the details of
someone to look up by CRN.
If a CRN is present in the flash then we request the details of that
person from the personService and present these details for the user to
confirm before starting the ‘Apply’ form