Make ‘back’ button on releaseDate page dynamic
Now there are multiple possible routes through the form having a hard
coded value for the back button doesn’t work. Some users might go through
pages A > B > D whilst others might go A > C > D. Therefore we can’t
always say ‘redirect to B’ because some users might have come through
page C instead.
Initially I thought we could determine which page to direct the user to
based on whether they had answers to a certain question in their session.
EG if a user had an answer to B then we could redirect them to B as they
won’t have seen page C but if the user went back to an early enough
point and answered questions differently then they might have answers
for both C and D in their session and we don’t know which one they answered
most recently.
This called for a different approach. When saving a response to a question
we now add a property ‘previousPage’ to the flash with the title of the page
whose response is being saved.
When constructing the Page object in the `getCurrentPage` method in the
`applicationService` we retrieve the ‘previousPage’ property from the flash
and pass it as a third property to the Page object’s constructor.
Currently this parameter is only used by the releaseDate page but it could
be used by other pages if needed in the future.