Update validation to handle errors with JSONpath
This also tidies up the way we fetch the translations for an error by
first calling a `generateErrors` method which returns an object with
a key of the field that caused the error and a value of the error
message, we can then call `generateErrorMessages` and
`generateErrorSummary` using the generated error object, rather than
having to do lookups on the imported `errorLookup` JSON more than once.

I’ve also added some error handling for the error lookups, which should
hopefully make debugging easier in the future.