Add Renovate config
We currently use Dependabot for managing updates, but it seems that a
lot of teams are migrating to Renovate now, so I thought we’d give it a
go. We get a lot more control over what dependencies we can update, and
can be a lot smarter about rebasing etc.