Updating an application handles null data
When an application is first returned from the API, the `data` attribute
is `null`, so we need to initialize `data` with an empty object. I’ve
updated the tests to be more inline with reality and also added another
unit test to make sure the session data gets updated correctly.