Remove docker container names when they aren't required
Setting names can conflict with other projects, such as the API.

When we remove the name, docker compose will use the directory name as a prefix, ensuring uniqueness: `hmpps-approved-premises-api-2_hmpps-auth_1` rather than `hmpps-auth_1`.

This removes a conflict with wiremock with the API and allows both projects to be started independently.