Fix need to bootstrap on each server start
After the prior change to the script/server to tear down containers when the process was ended, we uncovered a problem when using the app with wiremock turned on.

It seems that removing containers removes some state that's important to wiremock. We were having to run script/bootstrap each time. On taking a look we see that there's some stubbing of API endpoints, presumably via Wiremock(?). As wiremock is a backing service and a dependency to have set on each server start we move it to the start-backing-services script so that it's called by both bootstrap and server scripts.