Add scripts
This script pulls from the types from the api spec (https://github.com/ministryofjustice/hmpps-approved-premises-api/blob/main/src/main/resources/static/api.yml)
and generates typescript types for it which it dumps in the @types folder.
This is almost perfect but unfortunately we need the index file to be in a
*.d.ts format not *.ts so we have to rename it afterwards.
Rather than muddying up the package.json I have opted to extract this to the
`script` folder.
I have also added this script to the `test` script so we are using the most
recent types when we typecheck in CI