Update Error typing in the TasklistPage
The errors are now a record type where the only permitted keys are the
keys in a page’s body. This ensures type safety when we build up the
errors.

We also add a generic to the `ValidationError`, so the constructor knows
what type of errors to expect.

I’ve also made all the methods in the abstract class `abstract` too,
insead of optional, as all pages should implement them.