App can deploy into the new CAS namespace
The new community-accommodation-dev namespace is being defined here [1]. This change updates the existing Approved Premises configuration that we copied across and gets temporary-accommodation ready to launch into the new namespace.

Whilst the API service is referenced, it is not deployed into this new namespace yet so some cooridination will be needed.

We found that this line `tlsSecretName: temporary-accommodation-ui-cert # override per environment` includes a value that we can't trace. We think it's a default that if it were used would cause issues, however in values-dev.yml this key is being set to `hmpps-temporary-accommodation-dev-cert` which is known. We aren't sure if we can remove the former without causing an issue so we leave it the same as AP for now.

[1] https://github.com/ministryofjustice/cloud-platform-environments/pull/9081