Explicitly set a premises ID to "unknown" in our premises E2E tests
In our E2E tests, we are creating a premises to correspond to a premises in the database, without knowing what its ID will be. Rather than using the autogenerated ID from the factory, we set it to "unknown", to aid any future debugging. We will follow this convention in future E2E tests