Enable E2E tests
We copy the original Approved Premises approach that runs the e2e only tests after deployment.

Ideally we'd have an ephemeral environment we could deploy and test against first, before deploying to real environments. Support for ephemeral environments isn't something we're aware of.

Right now main is deploying into the dev (staging) environment so no real users would be affected by this. As we add environments we'll likely have to revist this.

For now, if we deploy a change to the dev environment that ends up breaking the e2e tests we will be notified and be able to fix it.