Allow the user to set a premises status when creating or editing a premises
We add radio buttons to our template, with the full list of status supplied by PremisesController.

In our controller tests, we revise how we exercise the behaviour that an empty list of characteristic IDs will be supplied to the API, if none are supplied from the form. (This is needed as the API expects a list, but browsers will omit the field entirely rather than sending an empty list). Rather than creating a request body with a small subset of a NewPremises, we use a "full" NewPremises, but with characteristicIds deleted