Add insertGenericError helper function
We will use this in cases where we want to handler API errors that don't follow the format of giving an "errorType" for a "propertyName". This will allow us to handle these errors using the same conventions as other errors. We will use this when the API returns a 409 Conflict error, which will occur if we try to create conflicting bookings