Route BookingReportsController
We make BookingReportsController's methods reachable by URL. This allows us to add integration testing to booking report downloads.

Unusually, we test that we see an error page when the API returns an error, even though there are no expected errors. This is to test that error responses from the API do not simply get piped into a downloaded file

Update report.cy.ts