Update DashboardController
We update DashboardController to present a dashboard page, rather than redirect to our premises listing. The CSS for this page presents a one, two, or three column set of cards, depending on screen size.

We also add a DashboardPage helper class, to help us work with this page in our integration tests. We need here to update our login E2E tests, as the user will land on the dashboard, rather than the premises listing