Use node polyfill in E2E tests
We will want to use the "path" module in our E2E tests, when constructing the path to inspect to verify file downloads. To support this we need to add NodeModulesPolyfillPlugin to our preprocessor. We do not need to do this for our integration tests, where "path" just works, probably because those do not use Cucumber