Only show latest extension notes
If a booking has been extended, only show the latest extension notes, and only show if the booking is in the arrived state. Since we will be displaying the booking history, showing all extension notes always is not necessary. We use the getLatestExtension helper we added in the previous commit