Reflect PDU being a required field
We setup our wiremocked API to reflect the real API by requiring that PDU be set. We also test in our E2E tests that an appropriate error is show if the PDU is not set. The presence of the PDU is not checked by the API unless everything else is present, hence the sightly awkward setup of the E2E tests and the wiremocked API.

In our integration tests, we're mainly testing that we're showing the correct error, and that this error string is present, so we assume the presence of all require fields is tested for in one step for the sake of brevity