Remove redundant dependency lines in CircleCI
The requirement for `helm_lint` and `build_docker` was effectively duplicated.

The third job to run is called `deploy_dev` which must always run. This job depends on the first two jobs: `helm_lint` and `build_docker`.

Therefore we can assume that any job thereafter that depends on `deploy_dev` can inherit this instead.

Though functionally it shouldn't change anything the circleci deployment pipeline graph will be simpler as a result.