Fix intermitent E2E test failures
Some of our E2E tests were occasionally failing due to cypress testing an unexpected element. For instance, we'd try to get the "Warwickshire" <option> element in a dropdown list, but we'd actually test the "North Warwickshire" element. We add a small helper function to ensure we match the exact string, rather than a substring