Add AuditService
Modelled after the AuditService in the make-recall-decision-ui project, we add an AuditService for adding messages to a configured AWS SQS queue.

For now, we do not require auditing to be configured, and we switch off logging. We will change this once the SQS Queue is set up by the Cloud Team, and the necessary environment variables have been set.