Check presence of user UUID in auditMiddleware
This adds an extra layer of safety, ensuring that a user UUID is always attached to routes we expect to be audited. The response of logging an error and redirecting to /authError is modelled after authorisationMiddleware.

We also at a UUID to our stubbed user to ensure our integration tests can run.