Pass request into RestClient
We modify our clients, services, and controllers, so we can get the request into the constructor of our RestClient. We need this as we will next want to include the user's region in the header of each request

As part of this, we add a new createMockRequest test util function. We cannot simply use createMock<Request>, as jest fails to match the returned object when passed into expect().toHaveBeenCalledWith() checks