Add user region to session
We add populateUserRegion middleware, and modify setUpCurrentUser, to add the user region to the session if not already present.

Additionally, we need to make changes to each of our integration tests to stub the API endpoint we use to get the user region. As part of this, we wrap each of the user and the probation region with an alias, as we will need to access these in future integration tests