Use MoJ page header bar
We replace the use of the MoJ identity bar component, with the more appropriate MoJ page header bar component. The necessitates some minor integration test changes, to allow Cypress to find the action buttons. We also need some changes in bookingUtils, as the action buttons are populated from a list of items, rather than a list of lists of items