populateCurrentUser passes CallConfig object
We update populateCurrentUser so that it uses extractCallConfig to pass a CallConfig into our UserService.  Though populateCurrentUser was getting the token from the response before and we are now getting it from the request, we can look at setUpAuthentication.ts to see these will be the same.

For now, UserService simply takes the token from this object, and pass it into our UserClient as before.