Add getUrlEncodedCypressEnv helper function
We will be urlencode-ing one of our environment variables due to it containing a comma, so we add this helper function to get its decoded value. We need to do this as Cypress does not support commas in environment variables specified on the command line. See https://github.com/cypress-io/cypress/issues/4543