Replace getAnyPremises with samplePremises
We replace getAnyPremises with samplePremises in our PremisesListPage class. This will assign a random selection of premises to an alias.

When testing that we are only displaying premises for the user's region, we will want this more general method for sampling premises, and we can use this in place of our existing getAnyPremises method.