Allow ARK/DOM1 IP addresses to access the service
At the moment our users are not using a MOJ VPN and so can't access production.

They could access the Create and Vary a license service which manages it's infrastructure similar to us. We can see there that they have a much larger allow list[1]

We can see from the Manage and Supervision team that they have previously checked if the DOM1 network IP addresses will enable most of their users[2], which was confirmed except for Greater Manchester.

The IP addresses that we are adding here are taken from the Delius allow list[3] on the basis that if the user can access Delius, they should also be able to access our service.

We may have to add more IP addresses over time as we learn we need more.

[1] https://github.com/ministryofjustice/create-and-vary-a-licence/blob/main/helm_deploy/create-and-vary-a-licence/values.yaml#L74
[2] https://mojdt.slack.com/archives/C027KSVD0F5/p1634220661003900
[3] https://github.com/ministryofjustice/hmpps-env-configs/blob/a222488e6e7a3f233f793b80197384b2136b5aee/common-prod/common.tfvars#L413