Add debug sentry endpoint
Our goal is to have confidence production errors are being tracking for this app. Currently the prod environment isn't appearing at all. No errors, no transactions, just some session info for the dev and test environment.

However, when we make the AP UI throw an error we can see errors come through!

We follow this sentry guide for node to force an error from the frontend to test the connection: https://docs.sentry.io/platforms/node/guides/expressc

This will confirm our only slight theory is that receiving an error forces the environment to be recognised.