Add confirm request handler to BookingsController
We add our new request handler, and associated Nunjucks template. The confirm page will hold the entered booking details in hidden fields, which will then be submitted to the 'create' endpoint

We also add a BookingConfirmPage helper class, ready for when we test this request handler. The booking constructor parameter is optional, as for now the user may reach this page without having entered any booking details.