Vary booking text depending on extension type
We use the helper function we added in the previous commit to vary the text we display to the user for the notes section of the booking info.

We update our tests to expect this variation. In our E2E tests, we now need to ensure that the 'previousDepartureDate' field on the extension we create within the test properly mirrors the extension the UI will get from the API.