Use standard location header when editing a bedspace
We now show the bedspace reference, the premies reference, and the premises address when editing a bedspace, using the standard location header we use elsewhere in the UI. This fixes the bug where we were displaying the bedspace reference but with the label "Property reference".

To test this, our BedspaceEditPage helper class now needs to take the premises in its constructor.