Populate 'premises' alias in E2E tests
In later E2E tests around bedspaces, we will want to know the details of the premises the bedspace is in. Currently, when using an existing premises, we populate the 'premises' alias, but this only contains the limited details visible on the listings page. We now more fully populate the premises with the details we can scrape from the show premises page