Enable Sentry Tracing
This should hopefully start sending some event/transaction data to Sentry where we currently have none.

We copy from Make a Recall decision[1] and follow the Sentry guidance[2] for setting the value to 1.0 in prod.

The sample rate should limit the amount of data in other environments preventing the Sentry account from running up to it's transaction limits.

[1] https://github.com/ministryofjustice/manage-recalls-ui/blob/c3b4fd4486b26da47c086778fdc89e7a6befcde0/server/middleware/setUpSentry.ts
[2] https://docs.sentry.io/platforms/node/#configure