Add a location-header macro
We add a macro for a component that displays a premises reference, a bedspace reference, and a CRN, with each being optional. We will use this everywhere we want to display contextual information to the user for which premises/bedspace/CRN they are currently operating on. We support a 'hideAddress' parameter for the special case of the premises editing screen, where we don't want to display the address as it may be edited below by the user.

Concurrently, we add a helper for testing this component.