Load build info from correct location in order to fix Sentry
Sentry isn't being passed the source maps and so is unable to present meaningful stacke traces when errors occur.

During the build the build-info.json file is populated. Data here can then be accessed and forwarded to Sentry as we are already configured[1].

We copy the implementation made in AP [2][3].

[1] https://github.com/ministryofjustice/hmpps-temporary-accommodation-ui/blob/47e261d69589ed6419afb576dec0c35b67bb5157/server/middleware/setUpSentry.ts#L17
[2] https://github.com/ministryofjustice/hmpps-approved-premises-ui/pull/588
[3] https://github.com/ministryofjustice/hmpps-approved-premises-ui/pull/590