Integrate OptionalOasysSections page into Apply journey
We set our new optional OASys sections page as the first page in our OASys import task, and for this page to then return to the dashboard. We also add the containing section to appear next after reasons for placement, adding our new page to the Apply journey.

This requires us to now update our integration tests to complete this page, using the helper class we added in the previous commit.