Prevent double click on form buttons
We add the preventDoubleClick attribute on form submission buttons across the service. While this cannot eliminate double-POSTing entirely, it can reduce its prevalence. We make an exception for the bookings report form button, as this does not take the user to a new page, and preventing the button here from being clicked more than once may be counter intuitive