Propagate errors with empty 'invalid-params'
We previously attempted to handle any API error with an 'invalid-params' array present. However, if this were empty, we would swallow the original API error, then throw a new error when attempting to render the empty errors array. This gave us the worst of both worlds - it didn't present a friendly error to the user, and it obscured the underlying error to developers.

We now check that the array is non-empty, and if it is empty, we simply allow the error to propagate.