Describe common ways to test
Whilst script/test takes care of all the testing steps for us it isn't often practical to run the whole thing each time as it takes a long time.

During dev we may want to run individual tests that are faster to get feedback from.

I couldn't figure out exactly what the purpose of the `npm run test:e2e` was as it doesn't seem to work currently, so I've left that out for now.