Hide "Add bedspace" action for archived premises
We add a premisesActions utility function, that will return the available actions, given a premises' status. The output of this replaces the action list previously hardcoded into the show premises Nunjucks template. This follows a similar pattern to how we display the correct actions for bookings.

We update our E2E/integration tests to check for the presence/absence of premises actions as appropriate.