Hide actions for archived bedspaces in archived premises
Similar to our change for premises, we add a bedspaceActions utility function, that will return the available actions, given the containing premises' status. The output of this replaces the action list previously hardcoded into the show premises Nunjucks template.

We update our E2E/integration tests to check for the presence/absence of bedspace actions as appropriate.