Update linting rules
We update our linting rules to not require files with a single export to use a default export. This means, for example, when we expand a new utility file from one function to two, we don't have to change how the first method is imported. This is also consistent with Approved Premises' linting rules