Tidy deriveBookingHistory helper
Ahead of expanding this function to support edited cancellations and departures, we change functions such as getPreviousDepartedBookingState so their exits are in the order we expect them to be reached when the function is called multiple times. This will simplify the code we will add later to this function.

We also extract a compareBookingState function, as we will want to compare cancellations and departures, as well as extensions.