Fix exact helper
We use this helper function to generate a regular expression to exactly match a string. This is useful in our integration tests, but for strings containing characters with special meaning within a regular expression, we were not generating the correct regular expression. We now fix this by escaping the input string before using it to construct the regular expression