Add PoP details header to booking pages
We add a PoP details header component that sits above the location details header on booking related pages (and remove the CRN from the location header).

We also add this header to the Confirm CRN page, which requires us to check the CRN on the previous ‘new booking’ page, and add 404 and 403 error handling to remain on the page.