Add BedspaceSearchResult component
We add a Nunjucks component for displaying a single bedspace search result.

Our helper for testing this component in integration and E2E tests takes an optional checkCount parameter. When E2E testing, we will not be able to guarantee the number of bedspaces in the property. We will be working with pre-existing data, so we cannot safely test this.