Add BedspaceSearchController
Our bedspace search feature is implemented as a single route, that will search bedspaces using query data arriving from the search form. If no query data is provided (as will be the case when the user is arriving on the search page for the first time), the form will be presented without search results.

Again, our integration and E2E test helper takes an optional checkCount parameter. This is needed as we cannot guarantee the number of search results in our E2E tests, so we cannot safely test this in this context.