Update RestClient
We update our REST client so that it no longer filters out zero values, and so that it replaces Number.NaN (which cannot be serialized to JSON) with 'not-a-number'.

If the API receives a string when it is expecting a number, it will return a 'expectNumber' error. In combination with parseNaturalNumber returning NaN for non-numeric values, this means we will get a helpful error from the API if the user enters a non-numeric value where a number is expected.