Update Premises factory
In our E2E tests, we want to move away from using a set of reference data known by the UI (which is prone to get out of sync with the API) and towards using data derived from the page the E2E test is interacting with. We add a forEnvironment method to create a premises from this data we will read from a page

As this new helpers is only used in E2E tests, we need to mark it to be ignored by our test coverage checker.