Update UpdatePremises and NewPremises factories
A common pattern in our E2E tests is to building an object such as a Premises (as would be received from the API), and then using this to derive an object such as NewPremises or an UpdatePremises (as would be sent to the API). For premises, we add some helper functions to our factories to assist in following this pattern.

As these new helpers are only used in E2E tests, we need to mark them to be ignored by our test coverage checker.