Get PDUs from reference data in PremisesService
We update PremisesService to retrieve PDUs via the ReferenceDataClient, rather than using our JSON file. Note that for now we remap the PDU ID to be its name - the UI is still sending and receiving the PDU name to and from the API when editing, creating, and viewing a premises, so we treat the name as the identifier.