Get PDUs from reference data in BedspaceSearchService
We update BedspaceSearchService to retrieve PDUs via the ReferenceDataClient, rather than using our JSON file. This requires us to make some changes to our integration tests, to ensure the required reference data is stubbed.

Note that for now we remap the PDU ID to be its name - the UI is still sending the PDU name to the API searching, so we treat the name as the identifier.