Populate probationDeliveryUnit in PremisesShowPage
We update getPremises in PremisesShowPage to populate probationDeliveryUnit in the premises object derived from the page contents.

Note that it uses the name of the PDU as its ID. This means the ID will be inaccurate, but this will work for the purposes of E2E tests. Ultimately, the ID will be used to find an entry in a dropdown list, using Cypress's select method. This takes a string, in our case the PDU ID, but will compare it to both the value (corresponding to a PDU ID) and the user-facing text of each dropdown option, hence it will not matter in practice that the ID of the PDU is set to its name