Change PDU name handling for E2E tests
We change the PDU name checking to ensure that an API PDU name (e.g. 'North West Lancashire (Blackpool and lower tier LAs in Lancashire - Fylde, Wyre and Lancaster)' starts with the UI substring (e.g. 'North West Lancashire'), rather than checking for an exact match. This is because some of the PDU names stored in the UI are shortened versions of those stored in the API, which has caused E2E tests to fail.