Dev environment automatically checks for very new node versions
We found that Renovate proposed we update node 1 minor version from 18.15.0 to 18.16.0. 4 days later the latest version of node-build failed to contain this new version which was eventually added[1]. This meant that attempts to run `script/setup` failed due to node 18.16.0 not being available local and attempts to install it failed since it wasn't an available version, at least via node-build.

CAS1 linked us to this nodenv package[2] that seems to account for this situation and allows our machine to build new definitions without delay.

We want Renovate to keep proposeing minor and patch updates quickly in order to keep the service secure. This means that we need to accound for the fact that our local environments may again fall behind. Given  the `script/setup` is run after new updates (and we advise returning devs to run it) and that has a dependant `script/bootstrap` script, this seems like the best place to do it.

[1] https://github.com/nodenv/node-build/blob/master/share/node-build/18.16.0
[2] https://github.com/nodenv/node-build-update-defs.git